fix(os): resolve triple faults and VMware stability issues with interrupt hardening
This commit implements comprehensive fixes for VMware stability and triple fault issues that have plagued the GamerOS kernel. The changes include multiple layers of defensive programming across interrupt handling, device drivers, and system initialization.

Key fixes include: IDT hardening ensuring unset vectors are properly marked as not-present gates; ISR stack alignment fixes maintaining SysV ABI compliance for C function calls; exception handling improvements that prevent recursive exception scenarios from escalating to triple faults; keyboard driver fixes filtering out auxiliary device (mouse) bytes that were previously misinterpreted as scancodes; and mouse packet overflow protection to prevent large coordinate jumps.

Additionally, this commit marks 11 significant bugs as resolved: VESA mode functions, IPC system implementation, executive services initialization, Notepad text editing, real-time clock, start menu, memory leak in kfree coalescing, object manager memory allocation, user mode subsystems, UI framework implementation, window manager operations, and I/O manager async operations.

The executive layer now properly initializes GDI, registers input drivers, and creates baseline filesystem objects. The taskbar clock now reads from RTC instead of displaying static time. A compact font rendering system improves UI readability, and polling-only input mode provides fallback stability when IRQ-driven paths cause issues in VMware.

+++ src/intf/stdint.h
@@ -4,16 +4,16 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 
 typedef signed char int8_t;
 typedef signed short int16_t;
 typedef signed int int32_t;
-typedef signed long int64_t;
+typedef signed long long int64_t;
 
-typedef unsigned long size_t;
-typedef unsigned long uintptr_t;
-typedef signed long intptr_t;
+typedef uint64_t size_t;
+typedef uint64_t uintptr_t;
+typedef int64_t intptr_t;
 
 #ifndef NULL
 #define NULL ((void*)0)
